requirements.txt
tqdm >= 4.23.3
wget >= 3.2
matplotlib
Pillow
pydotplus
scikit-learn
scipy